Replace GitHub matrix test with a for loop in a test script.
The matrix test creates a new container for each version, which means that Bazel
can’t cache any outputs between the versions.  While a script with a loop is
less elegant, it’s significantly faster and cheaper overall.

In the script, dynamically select supported Emacs versions/toolchains based on
the operating system so that the script can also run on macOS.