Support Windows paths in compilation-mode (#314)
* Support Windows paths in compilation-mode

The regexps added to `compilation-error-regexp-alist-alist` need to match paths
output by Bazel, which are subject to numerous constraints.  As currently
written, this does not support Windows paths, which, as output by Bazel, start
with a drive letter and a root path (e.g `C:/`).

Address this by extending the regexp to account for Windows paths.  UNIX-style
path processing should not be affected; tests passed there.

NOTE: I was not able to build the test suite on Windows, nor do there seem to be
instructions for doing so.  I'll be happy to discuss what I tried in the
PR, but I'm not exactly a Windows expert (nor do I do development on it regularly).