Simplify interface of ‘bazel--target-completion-table’ a bit. (#312)
Instead of requiring the caller to pass in workspace root and package name,
derived them from ‘default-directory’.  This changes the semantics for some
exotic cases (e.g. when the directory of ‘buffer-file-name’ is not the same as
‘default-directory’), but that hopefully doesn’t matter in practice.  It
simplifies the callers a bit.

This also serves as preparation for fixing the behavior in directories which are
not packages.  The actual fix for that will come as a follow-up PR.